// DashBoardAngular/src/app/content/rack-mapping/models/ressourceHardwareDetail.model.ts
export class RessourceHardwareDetail {
    ID_RessourceHardware: number;
    Brand: string;
    Model: string;
    SerialNumber: string;
    Price: number;
    Owner: string;
    Status: string;
    Origin: string;
    Category: string;
    Type: string;
    TechnicalInfo: string;
    Comments: string;
    DateIn: string;
    DateImport: string;
    BarCode: string;
    Configuration: string;
    ModuleCommand: string;
    ModuleSpecific: string;
}